First implementation.  Native and numpy ufuncs.
The native implementation is numpy-agnostic and defines
the graph classes, ``Priop`` and ``GraphEdge``.  A couple
of changes have been made with respect to the initial
proposal.

There is also a numpy specific wrapper for ufuncs using
the ``Priop`` class.  Importing that module installs the
wrappers in numpy using ``numpy.set_numeric_ops()``.